changed how conversations are represented, added message loading
-Conversations are the things that are going to be selected in the 'Main
Fragment' so I rewrote most of the areas which said 'currentUser' and
replaced it with 'currentConversation'
-removed limits on the login mechanism, like email needs an @
-store the current user and auth token in shared preferences.  If the
Main Activity detects theres no auth token, it'll send you back to the
login page
-split up the onUserClick to onConversation click, which is implemented,
and the staff subsection click, which isnt.
-reworked the HTTP calls, and added 'on complete' listeners.  I don't
know why, but doing any HTTP calls on the main thread, which it seems
like you're doing, should throw errors.  If the error does come up,
it'll be easy to fix it now? hopefully?
-removed the randomness of the 'chat bubbles' now.  if you said it,
it'll be on the right.
-inputting messages has been disabled now.  it needs to be properly
implemented.
-changed the way tags work.  Now you'll be able to tell if the ID you're
getting is a conversation tag, from a conversation, or a user tag, from
the staff click
-have to wrap those calls to run on the main ui thread, or exceptions
will be thrown (navdrawer adapter)
-removed some public declarations since androidstudio kept showing those
'errors/suggestions'
-subclassed user and conversation to 'navdraweritem' and added/moved
appropriate methods to it
-made all the sessionwrapper calls static, but left the logic of the
string parsing in there for you to do as you want.
-changed 'onUserClickedOn' to 'onViewClicked' though I've never heard of
putting your 'on click methods' in the XML,  probably separation
concerns